requirements.txt
coloredlogs==10.0
eventlet==0.24.1
gunicorn==19.9.0
SQLAlchemy==1.2.17
biopython==1.73
bottle==0.12.16
ete3==3.1.1
matplotlib==3.0.2
numpy==1.16.1
pandas==0.24.1
PyMySQL==0.9.3
seaborn==0.9.0
scipy==1.2.0
six==1.11.0
